🤖 Refactor(client): 엣지 분할 코드 hook으로 분리
